sendermode: refactor
